+++ util/encoding.go
@@ -6,16 +6,16 @@ var (
 	maxVarintLength = 5
 )
 
-func Varint(n int64) []byte {
+func Varint(n int32) []byte {
 	buf := make([]byte, maxVarintLength)
-	l := binary.PutVarint(buf, n)
+	l := binary.PutVarint(buf, int64(n))
 
 	return buf[:l]
 }
 
-func Uvarint(n uint64) []byte {
+func Uvarint(n uint32) []byte {
 	buf := make([]byte, maxVarintLength)
-	l := binary.PutUvarint(buf, n)
+	l := binary.PutUvarint(buf, uint64(n))
 
 	return buf[:l]
 }
